+++ src/command/command.service.ts
@@ -68,6 +68,10 @@ export class CommandService {
     void this.setStoredCommands();
   }
 
+  static isCommandFormat(message: string) {
+    return /^![a-z0-9]+/i.test(message);
+  }
+
   setInitialCommandState() {
     this.commandState = {
       arena: '',