+++ src/command/commands/wordle.ts
@@ -15,7 +15,7 @@ const command: Command = {
   name: 'wordle',
   help: 'Guess a 5-letter word by typing !wordle <word>. You may only guess once every 60 seconds. Capital means correct space. Lowercase means wrong space. If the word is funky and you guess frown: F---n.',
   platforms: [Platform.Twitch, Platform.Discord],
-  run: async (ctx) => {
+  run: async (ctx, { services }) => {
     const guess = ctx.body?.toLowerCase().trim();
 
     if (guess === 'new' && ctx.isOwner) {
@@ -74,6 +74,7 @@ const command: Command = {
       ctx.botSpeak(
         `@${ctx.displayName} got it! The word was ${chosenWord.toUpperCase()}!`
       );
+      void services.twitchService.ownerRunCommand(`!define ${chosenWord}`);
       chosenWord = undefined;
       users = {};
       return true;