+++ RELEASE_NOTES.md
@@ -17,6 +17,7 @@
 - Fixed a bug in `!opp` and `!king` where a user could change to a secret king.
 - Now catching errors in `!image` creation and letting the user know their request failed.
 - Make chat `!personality` clearer in the prompt.
+- Fixed a bug where `!!` and `! ` were being treated as commands. Now only `!` followed by at least 1 alphanumeric character is treated as a command. Thanks @dannovikov @fitztrev!
 
 ### 1.10.1 (2024-02-05)